fix(audit): pass 6 — restore pruned bake-result include, HDR clamp removal, guard-pinning test, sibling routing on attach bail
